+++ Back-End/app/Http/Controllers/ServiceOrderController.php
@@ -17,30 +17,41 @@ class ServiceOrderController extends Controller
         return response()->json(ServiceOrder::all());
     }
 
+
+    public function OrderClient()
+    {
+        $user = auth()->user();
+        $clientId = $user->client->id;
+        $orders = ServiceOrder::where('client_id', $clientId)->get();
+        return response()->json($orders);
+    }
+
     /**
      * Show the form for creating a new resource.
      */
-    public function create(Request $request,Service $service,Client $client)
+    public function create(Request $request, Service $service)
     {
-        $data=$request->validate([
-            'order_date'=>'required|date',
-            'deliverr_date'=>'required|date',
-            'total_amount'=>'required|numeric',
+        $data = $request->validate([
+            'order_date' => 'required|date',
+            'deliverr_date' => 'required|date',
+            'total_amount' => 'required|numeric',
 
         ]);
+        $user = auth()->user();
+        $data['client_id'] = $user->client->id;
         ServiceOrder::create($data);
-        
-        return response()->json(['message'=>'Order sent Successfuly','Order'=>$data]);
+
+        return response()->json(['message' => 'Order sent Successfuly', 'Order' => $data]);
     }
 
-    
+
 
     /**
      * Display the specified resource.
      */
     public function show(ServiceOrder $serviceOrder)
     {
-        return response()->json( ServiceOrder::find($serviceOrder));
+        return response()->json(ServiceOrder::find($serviceOrder));
     }
 
 
@@ -49,8 +60,12 @@ class ServiceOrderController extends Controller
      */
     public function destroy(ServiceOrder $serviceOrder)
     {
-        ServiceOrder::findOrFail($serviceOrder);
-        $serviceOrder->delete();
-        return response()->json(['meesage'=>'Order delete Sccessfully']);
+        if ($serviceOrder->status === 'complete') {
+            return response()->json(['message' => 'you can not delete ']);
+        } else {
+            ServiceOrder::findOrFail($serviceOrder);
+            $serviceOrder->delete();
+            return response()->json(['meesage' => 'Order delete Sccessfully']);
+        }
     }
 }
